Correct Date indexing for pass-by-ref DataFrame
There was an issue where two functions were setting the index to the
Date column.  Both of these functions received a DataFrame from a
calling function.  The updating of the DataFrame in-place when setting
the Date index on the first call changed the object for subsequent
calls. This was repaired by making a copy.